api/requirements.txt
fastapi
web3
python-decouple
deta
